made side bar own component
